Moving stuff about dialogs in a separate module nurse.game.dialog
Added the nurse.game.dialog module (and removed from sprite) and
example dialog.py
Corrected backends/__init__ and sprite with new path of dialog module
The example challenge1 was corrected as well in that sense.
