+++ nurse/sprite.py
@@ -322,91 +322,6 @@ class UniformLayer(Sprite):
 		pass
 
 
-class DialogState(State):
-	
-	def _parse_lines(self):
-		import string
-		words = string.split(self._text, ' ')
-		current_text = words[0]
-		lines = [ current_text ]	
-		
-		for word in words[1:] :
-			current_text = string.join([current_text, word], ' ')
-			repr = Config.get_graphic_engine().load_text(\
-				current_text, self.font, self.font_size,
-				0,0)
-			if repr.content_width > self.max_width:
-				current_text = word
-				lines.append(word)
-			else:
-				lines[-1] = current_text
-		return lines
-			
-	
-	def __init__(self, name='text', text='...', font='Times New Roman',
-			font_size=20, text_area=None, perso=None, char_per_sec=5.,
-			writing_machine_mode=True):
-		State.__init__(self,name)
-		self._text = text
-		self.font = font
-		self.font_size = font_size
-		self.max_width = text_area[1][0]
-		self._lines = self._parse_lines()
-		self.max_lines = 5
-		self.perso = perso
-		if writing_machine_mode:
-			self.char_delay = 1000. / char_per_sec
-		else:	self.char_delay = 0
-		self.writing_machine_mode = writing_machine_mode
-		self.list_backend_repr = []
-		self._current_time = 0
-		self._current_indice = 0
-		self._current_text = ''
-		self._current_line = 0
-		self._current_height = 0
-		self.emit('dialog_state_started')
-		print 'dialog_state_started'
-	
-	def _update_chars(self, n):
-		max_ind = len(self._lines[self._current_line])
-		ind = self._current_indice
-		if ind == max_ind: 
-			if self._current_line == len(self._lines) - 1:
-				self.emit('dialog_state_terminated')
-				self._fsm.stop()
-				print 'dialog_state_terminated'
-				return
-			ind = 0
-			self._current_text = ''
-			self._current_line += 1
-			self._current_height += (self._current_line % self.max_lines) * self.list_backend_repr[-1].content_height
-
-		new_ind = ind + n
-		if new_ind > max_ind: new_ind = max_ind
-		new_text = self._lines[self._current_line][ind:new_ind]
-		self._current_text += new_text
-		self._current_indice = new_ind
-		anchor_x, anchor_y = self._fsm.get_location()
-		repr = Config.get_graphic_engine().load_text(\
-				self._current_text, self.font, self.font_size,
-				anchor_x, anchor_y + self._current_height)
-		if len(self.list_backend_repr) == 0:
-			self.list_backend_repr.append(repr)
-		else:	self.list_backend_repr[-1] = repr
-		if len(self._current_text) == len(self._lines[self._current_line]):
-			self.list_backend_repr.append(repr)
-				
-	def update(self, dt):
-		self._current_time += dt	
-		if self._current_time >= self.char_delay:
-			if self.char_delay == 0:
-				n = len(self._text)
-			else:	n = int(self._current_time / self.char_delay)
-			self._update_chars(n)
-			self._current_time -= n * self.char_delay
-		
-
-
 class Dialog(Sprite):
 	def __init__(self, name='dialog', context=None, layer=2):
 		Sprite.__init__(self, name, context, layer)
